feat: more tests + other misc changes
